app/requirements.txt
pandas
streamlit
plotly
requests 
beautifulsoup4
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
apache-airflow-providers-standard
psycopg2-binary